+++ code/server/sv_world.c
@@ -732,9 +732,14 @@ SV_Trace(trace_t *results, const vec3_t start, const vec3_t mins, const vec3_t m
 
   Com_Memset(&clip, 0, sizeof(moveclip_t));
 
+  if ((unsigned)passEntityNum > MAX_GENTITIES - 1)
+  {
+    passEntityNum = ENTITYNUM_NONE;
+  }
+
   //clip to world
   CM_BoxTrace(&clip.trace, start, end, mins, maxs, 0, contentmask, type);
-  clip.trace.entityNum = clip.trace.fraction != 1.0f ? ENTITYNUM_WORLD:ENTITYNUM_NONE;
+  clip.trace.entityNum = clip.trace.fraction != 1.0 ? ENTITYNUM_WORLD:ENTITYNUM_NONE;
 
   if (!clip.trace.fraction)
   {